feat: add initial Kubernetes agent that creates a resource graph
Add a Kubernetes controller that watches various Kubernetes resources
and stores it in an in-memory KV-based store as a resource graph. The
graph defines relationships between Kubernetes resources.

The controller started off as a Kubebuilder/controler-runtime project
but ultimately went with managing the informers manually. At this point
we only use the controller-runtime Manager to orchestrate the controller
but this will likely be factored out as well and will strictly use
lower-level client-go APIs for the controller. We do this because this
is a not a typical controller where it needs to reconcile resources and
since resources are stored in the resource graph, have a less of a need
for controller-runtime to manage caching.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>