fix: make Store.UpdateResource be a replace rather than merge
proto.Merge semantics match the semantics of concatenating two byte
streams which isn't necessarily what we want. In particular, lists are
appended to the original list which meant we would get duplicate Tags on
resource updates.

So update Store.UpdateResource API to use replace semantics instead of
Merge. This can be revisted in the future but for now we always know
what the full updated Resource is so this is fine for now.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>