fix: properly shutdown intake worker
When the intake worker shutsdown or fails, the Manager initiates a
shutdown. When the worker shuts down, it stops consuming on
the resource store subscription channel. The Kubernetes Controller,
which is sending updates on this channel when it is updating the
resource store, is blocked because there is no consumer anymore. This
prevents the controller from properly shutting down and timing
out.

Have the intake worker begin shutting down only when the subscription
channel is closed. This streamlines the shutdown sequence.

SIGTERM -> [controller,resource store exits] -> signals intake
worker to exit

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>