feat: add TTL, delta version, and heartbeats when sending to intake svc
Generate a random 4 byte hex encoded string upon startup that serves as
the delta version that will be applied to every object sent to the
intake service. A new delta version is create between System Agent start
ups.

Add a TTL of 10 minutes to every object before sending to intake
service.

Send a Heartbeat delta to the intake service every minute. This extends
the TTL for every object a particular instance of the System Agent has
sent.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>