refactor: create global registry for collector factories
The registry is responsible for providing factories to create a
collector for a metric type. The factory takes in a config which will
eventually be supplied by the Manager which will maintain collector
configs.

Each collector registers itself to the registry using init(). Note that
only 1 collector can be registered per metric type - it will panic if
multiple collectors for the same metric type are attempting to register.

The registry also only contains ContinuousCollectors so that the manager
has only 1 Collector interface do deal with and orchestrate. New helper
methods PartialNewContinuousPointCollector and
PartialNewOnceContinuousCollector are added to transform PointCollectors
to ContinuousCollectors and add them to the registry.

Also update CLAUDE.md to add new collectors to the registry during
collector development and deal with some minor lint errors.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>